Update LibVLC to 3.1.7 for Android, clean up and document Flutter code
