+++ .idea/runConfigurations/example_lib_main_dart.xml
@@ -1,6 +1,6 @@
 <component name="ProjectRunConfigurationManager">
-  <configuration default="false" name="example/lib/main.dart" type="FlutterRunConfigurationType" factoryName="Flutter">
+  <configuration default="false" name="example/lib/main.dart" type="FlutterRunConfigurationType" factoryName="Flutter" singleton="false">
     <option name="filePath" value="$PROJECT_DIR$/example/lib/main.dart" />
-    <method />
+    <method v="2" />
   </configuration>
 </component>
