+++ .idea/vcs.xml
@@ -2,5 +2,7 @@
 <project version="4">
   <component name="VcsDirectoryMappings">
     <mapping directory="$PROJECT_DIR$" vcs="Git" />
+    <mapping directory="$PROJECT_DIR$/android/vlc-android" vcs="Git" />
+    <mapping directory="$PROJECT_DIR$/android/vlc-android/vlc" vcs="Git" />
   </component>
 </project>
